+++ streamer/r2-streamer-swift/Parser/EPUB/NCXParser.swift
@@ -11,6 +11,7 @@
 
 import R2Shared
 import Fuzi
+import Foundation
 
 /// From IDPF a11y-guidelines content/nav/toc.html :
 /// "The NCX file is allowed for forwards compatibility purposes only. An EPUB 2