+++ streamer/r2-streamer-swiftTests/Asserts.swift
@@ -12,6 +12,11 @@
 import XCTest
 
 func AssertJSONEqual(_ json1: Any?, _ json2: Any?, file: StaticString = #file, line: UInt = #line) {
+    guard #available(iOS 11.0, *) else {
+        XCTFail("iOS 11 is required to run JSON tests")
+        return
+    }
+    
     guard let j1 = json1, let j2 = json2 else {
         if json1 != nil || json2 != nil {
             XCTFail("JSONs are not equal")