+++ streamer/r2-streamer-swiftTests/Fixtures.swift
@@ -6,6 +6,12 @@
 
 import Foundation
 
+#if !SWIFT_PACKAGE
+extension Bundle {
+    static let module = Bundle(for: Fixtures.self)
+}
+#endif
+
 class Fixtures {
     
     let path: String?
@@ -14,10 +20,8 @@ class Fixtures {
         self.path = path
     }
     
-    private lazy var bundle = Bundle(for: type(of: self))
-    
     func url(for filepath: String) -> URL {
-        return bundle.resourceURL!.appendingPathComponent("Fixtures/\(path ?? "")/\(filepath)")
+        return Bundle.module.resourceURL!.appendingPathComponent("Fixtures/\(path ?? "")/\(filepath)")
     }
 
     func data(at filepath: String) -> Data {